# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jewelry Store REST API",
    "description": "REST API theo database.sql 25 bảng. Đặt hàng COD; phân trang bắt đầu từ 0. KHACH_HANG, NHAN_VIEN, QUAN_LY. Xem README/API.md để biết phân quyền và luồng nghiệp vụ.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/wishlist/{id}": {
      "get": {
        "tags": [
          "wishlist-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "contains",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      },
      "put": {
        "tags": [
          "wishlist-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "add",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      },
      "delete": {
        "tags": [
          "wishlist-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
...
    "/api/wishlist": {
      "get": {
        "tags": [
          "wishlist-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "list_9",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "VariantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "reserved": {
            "type": "integer",
            "format": "int32"
          },
          "available": {
            "type": "integer",
            "format": "int32"
          },
          "salePrice": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DANG_BAN",
              "NGUNG_BAN"
            ]
          }
        }
      },
...
      "ImageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string"
          },
          "primaryImage": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "string"
          },
          "brandId": {
            "type": "integer",
            "format": "int64"
          },
          "brand": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "material": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "gemstone": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DANG_BAN",
              "NGUNG_BAN"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageResponse"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantResponse"
            }
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "reviews": {
            "type": "integer",
            "format": "int64"
          },
          "salePrice": {
            "type": "number"
          }
        }
      },
...
      "ApiResponsePageResponseProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseProductResponse"
          }
        }
      },
      "PageResponseProductResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}